modified character evaluation datatype

+++ Core/models.py
@@ -81,12 +81,12 @@ class CharacterEvaluation(models.Model):
     employee = models.ForeignKey(Employee, on_delete=models.CASCADE, related_name='evaluations')
     evaluator = models.ForeignKey(Employee, on_delete=models.CASCADE, related_name='given_evaluations')
     evaluation_date = models.DateField()
-    behavior_one = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
-    behavior_two = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
-    behavior_three = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
-    behavior_four = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
-    behavior_five = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
-    behavior_six = models.IntegerField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_one = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_two = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_three = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_four = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_five = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
+    behavior_six = models.FloatField(default=1,validators=[MinValueValidator(1), MaxValueValidator(4)])
 
     def __str__(self):
         return f'Evaluation for {self.employee.first_name} by {self.evaluator.last_name}'
